simplify code for badge managers
git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.products.badges/trunk@40870 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/nti/app/products/badges/subscribers.py
@@ -11,24 +11,25 @@ logger = __import__('logging').getLogger(__name__)
 from zope import component
 from zope.lifecycleevent.interfaces import IObjectRemovedEvent
 
+from nti.badges import interfaces as badge_interfaces
 from nti.badges.tahrir import interfaces as tahrir_interfaces
 
 from nti.dataserver import interfaces as nti_interfaces
 
 from nti.processlifetime import IAfterDatabaseOpenedEvent
 
-from . import get_user_badge_managers
-
 import sqlalchemy.exc
 
 @component.adapter(nti_interfaces.IUser, IObjectRemovedEvent)
 def _user_deleted(user, event):
-	for manager in get_user_badge_managers(user):
+	manager = component.getUtility(badge_interfaces.IBadgeManager)
+	if manager.person_exists(user):
 		manager.delete_person(user)
 
 @component.adapter(IAfterDatabaseOpenedEvent)
 def _after_database_opened_listener(event):
 	logger.info("Adding registered tahrir issuers")
+
 	import transaction
 	with transaction.manager:
 		# TODO: Should probably defer this until needed
@@ -38,30 +39,29 @@ def _after_database_opened_listener(event):
 		# database shard. We have some hacky defense against that below.
 		# It's also fired for every shard in every test case...not ideal
 		issuers = {x[1] for x in component.getUtilitiesFor(tahrir_interfaces.IIssuer)}
-		managers = {x[1] for x in component.getUtilitiesFor(tahrir_interfaces.ITahrirBadgeManager)}
-		for manager in managers:
-			if getattr(manager, '_v_installed', False):
-				continue
+		manager = component.getUtility(badge_interfaces.IBadgeManager)
+		if getattr(manager, '_v_installed', False):
+			return
 
-			setattr(manager, str('_v_installed'), True)
-			for issuer in issuers:
-				if not manager.issuer_exists(issuer):
-					# FIXME: Under some circumstances, we can get an
-					# IntegrityError: ConstraintViolation, even though
-					# this code path only checks name and origin.
-					# So clearly there's some sort of race condition here.
-					# Is our transaction not actually isolated? Or at the wrong level?
-					try:
-						manager.add_issuer(issuer)
-					except sqlalchemy.exc.IntegrityError:
-						logger.warn("Integrity error", exc_info=True)
-					else:
-						logger.debug("Issuer (%s,%s) added", issuer.name, issuer.origin)
-
+		setattr(manager, str('_v_installed'), True)
+		for issuer in issuers:
+			if not manager.issuer_exists(issuer):
+				# FIXME: Under some circumstances, we can get an
+				# IntegrityError: ConstraintViolation, even though
+				# this code path only checks name and origin.
+				# So clearly there's some sort of race condition here.
+				# Is our transaction not actually isolated? Or at the wrong level?
+				try:
+					manager.add_issuer(issuer)
+				except sqlalchemy.exc.IntegrityError:
+					logger.warn("Integrity error", exc_info=True)
+				else:
+					logger.debug("Issuer (%s,%s) added", issuer.name, issuer.origin)
 
 
-from nti.badges.openbadges.interfaces import IBadgeClass
 from nti.badges.tahrir.interfaces import IAssertion
+from nti.badges.openbadges.interfaces import IBadgeClass
+
 from zope.lifecycleevent import IObjectAddedEvent
 
 from .interfaces import SC_BADGE_EARNED
